fix install checks

+++ docker/dev/prepare_terrama2_base.sh
@@ -14,20 +14,27 @@ function valid()
 
 sudo apt-get update
 (
-  mkdir -p ${DEPENDENCIES_DIR}
-  cd ${DEPENDENCIES_DIR}
-  sudo apt install doxygen graphviz gnutls-bin gsasl libghc-gsasl-dev libgnutls-dev libssl-dev debhelper devscripts
-  wget https://github.com/kisli/vmime/archive/v0.9.2.tar.gz
-  cmake -G "Unix Makefiles" \
-        -DCMAKE_BUILD_TYPE:STRING="Release" \
-        -DVMIME_HAVE_MESSAGING_PROTO_SENDMAIL:BOOL=false \
-        -DVMIME_BUILD_SAMPLES:BOOL=false \
-        -DCMAKE_INSTALL_PREFIX:PATH=${DEPENDENCIES_DIR}
-  make -j4
-  make install
+  ls ${DEPENDENCIES_DIR}/lib/libvmime.so
+  if [[ $? -ne 0 ]]; then
+    mkdir -p ${DEPENDENCIES_DIR}
+    cd ${DEPENDENCIES_DIR}
+    sudo apt install doxygen graphviz gnutls-bin gsasl libghc-gsasl-dev libgnutls-dev libssl-dev debhelper devscripts
+    ls v0.9.2.tar.gz
+    if [[ $? -ne 0 ]]; then
+      wget https://github.com/kisli/vmime/archive/v0.9.2.tar.gz
+    fi
 
-  rm -rf v0.9.2*
-  rm -rf vmime*
+    cmake -G "Unix Makefiles" \
+          -DCMAKE_BUILD_TYPE:STRING="Release" \
+          -DVMIME_HAVE_MESSAGING_PROTO_SENDMAIL:BOOL=false \
+          -DVMIME_BUILD_SAMPLES:BOOL=false \
+          -DCMAKE_INSTALL_PREFIX:PATH=${DEPENDENCIES_DIR}
+    make -j4
+    make install
+
+    rm -rf v0.9.2*
+    rm -rf vmime*
+  fi
 )
 
 (
@@ -36,7 +43,7 @@ sudo apt-get update
   cd ${TERRAMA2_DIR}/codebase
   # check if terrama2 code is available
   git status
-  if [[ $1 -ne 0 ]]; then
+  if [[ $? -ne 0 ]]; then
     git clone https://github.com/TerraMA2/terrama2.git .
   fi
 )
@@ -46,7 +53,7 @@ sudo apt-get update
   cd ${TERRAMA2_DIR}/build-debug
   # check terrama2 build
   cmake .
-  if [[ $1 -ne 0 ]]; then
+  if [[ $? -ne 0 ]]; then
     cmake -G "CodeBlocks - Unix Makefiles" \
             -DCMAKE_BUILD_TYPE:STRING="Debug" \
             -DCMAKE_INSTALL_PREFIX:PATH=${DEPENDENCIES_DIR} \
